server_OIDC/requirements.txt
django
djangorestframework
django-oauth-toolkit
django-cors-middleware
django-auth-ldap